+++ app/Http/Controllers/CompatibilityRuleController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\StoreCompatibilityRuleRequest;
 use App\Models\CompatibilityRule;
 use App\Services\CompatibilityRuleService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CompatibilityRuleController extends Controller
 {
@@ -26,6 +27,7 @@ class CompatibilityRuleController extends Controller
      */
     public function store(StoreCompatibilityRuleRequest $request)
     {
+        Gate::authorize('create', CompatibilityRule::class);
         return $this->rule_service->create($request);
     }
 
@@ -34,6 +36,7 @@ class CompatibilityRuleController extends Controller
      */
     public function destroy(CompatibilityRule $rule)
     {
+        Gate::authorize('delete', $rule);
         return $this->rule_service->delete($rule);
     }
 }
